Move common widgets/utils to the shared folder
