+++ lib/models/node.dart
@@ -1,4 +1,4 @@
-import 'package:andax/utils.dart';
+import 'package:andax/shared/utils.dart';
 import 'transition.dart';
 
 enum EndingType { win, loss }
