+++ lib/models/story.dart
@@ -1,4 +1,4 @@
-import 'package:andax/utils.dart';
+import 'package:andax/shared/utils.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'actor.dart';
 import 'content_meta_data.dart';
