+++ lib/models/translation_asset.dart
@@ -1,6 +1,6 @@
 import 'package:andax/models/content_meta_data.dart';
 import 'package:andax/models/translation.dart';
-import 'package:andax/utils.dart';
+import 'package:andax/shared/utils.dart';
 import 'package:enum_to_string/enum_to_string.dart';
 
 enum AssetType {
