+++ lib/modules/home/screens/home.dart
@@ -5,13 +5,13 @@ import 'package:andax/modules/home/screens/search.dart';
 import 'package:andax/modules/story_info/screens/story_info.dart';
 import 'package:andax/store.dart';
 import 'package:andax/shared/extensions.dart';
-import 'package:andax/widgets/loading_builder.dart';
-import 'package:andax/widgets/paging_list.dart';
+import 'package:andax/shared/widgets/loading_builder.dart';
+import 'package:andax/shared/widgets/paging_list.dart';
 import 'package:andax/modules/home/widgets/raxys_logo.dart';
-import 'package:andax/widgets/rounded_back_button.dart';
-import 'package:andax/widgets/scrollable_modal_sheet.dart';
+import 'package:andax/shared/widgets/rounded_back_button.dart';
+import 'package:andax/shared/widgets/scrollable_modal_sheet.dart';
 import 'package:andax/modules/home/widgets/story_card.dart';
-import 'package:andax/widgets/story_tile.dart';
+import 'package:andax/shared/widgets/story_tile.dart';
 import 'package:andax/models/story.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
