+++ lib/modules/home/screens/search.dart
@@ -1,8 +1,8 @@
 import 'package:andax/modules/home/services/stories.dart';
 import 'package:andax/models/story.dart';
 import 'package:andax/modules/story_info/screens/story_info.dart';
-import 'package:andax/widgets/paging_list.dart';
-import 'package:andax/widgets/story_tile.dart';
+import 'package:andax/shared/widgets/paging_list.dart';
+import 'package:andax/shared/widgets/story_tile.dart';
 import 'package:flutter/material.dart';
 
 class SearchScreen extends StatefulWidget {
