+++ lib/modules/profile/screens/profile.dart
@@ -1,10 +1,10 @@
 import 'package:andax/models/story.dart';
 import 'package:andax/modules/story_info/screens/story_info.dart';
 import 'package:andax/store.dart';
-import 'package:andax/widgets/paging_list.dart';
-import 'package:andax/widgets/rounded_back_button.dart';
+import 'package:andax/shared/widgets/paging_list.dart';
+import 'package:andax/shared/widgets/rounded_back_button.dart';
 import 'package:andax/modules/profile/widgets/sign_in_buttons.dart';
-import 'package:andax/widgets/story_tile.dart';
+import 'package:andax/shared/widgets/story_tile.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
