+++ lib/modules/story_editor/screens/narrative_editor.dart
@@ -4,9 +4,9 @@ import 'package:andax/models/node.dart';
 import 'package:andax/models/story.dart';
 import 'package:andax/models/translation.dart';
 import 'package:andax/models/translation_asset.dart';
-import 'package:andax/widgets/loading_dialog.dart';
-import 'package:andax/widgets/maybe_pop_alert.dart';
-import 'package:andax/widgets/rounded_back_button.dart';
+import 'package:andax/shared/widgets/loading_dialog.dart';
+import 'package:andax/shared/widgets/maybe_pop_alert.dart';
+import 'package:andax/shared/widgets/rounded_back_button.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
