+++ lib/modules/story_editor/widgets/actor_picker.dart
@@ -1,7 +1,7 @@
 import 'package:andax/models/actor.dart';
 import 'package:andax/models/translation_asset.dart';
 import 'package:andax/modules/story_editor/widgets/modal_picker.dart';
-import 'package:andax/widgets/rounded_back_button.dart';
+import 'package:andax/shared/widgets/rounded_back_button.dart';
 import 'package:flutter/material.dart';
 
 import '../screens/narrative_editor.dart';
