+++ lib/modules/story_info/services/content_loader.dart
@@ -3,7 +3,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:andax/models/story.dart';
 import 'package:andax/models/translation_asset.dart';
-import 'package:andax/widgets/loading_dialog.dart';
+import 'package:andax/shared/widgets/loading_dialog.dart';
 
 Future<Translation> loadTranslation(StoryInfo info) async {
   final tdc = FirebaseFirestore.instance.doc(
