+++ lib/shared/widgets/snackbar_manager.dart
