+++ lib/shared/widgets/span_icon.dart
