+++ lib/shared/widgets/story_tile.dart
@@ -1,6 +1,6 @@
 import 'package:andax/models/story.dart';
-import 'package:andax/utils.dart';
-import 'package:andax/widgets/span_icon.dart';
+import 'package:andax/shared/utils.dart';
+import 'package:andax/shared/widgets/span_icon.dart';
 import 'package:flutter/material.dart';
 
 class StoryTile extends StatelessWidget {
